[SUREFIRE-1881] - Fix and extend integration test
- Renamed Surefire1881 to Surefire1881IT, because with the wrong name
  the test was not even executed before according to the CI build logs

- Now in addition to the Java agent logging IT executed in Failsafe,
  there also is a JVM logging IT executed in Surefire. Both tests were
  in different branches of
  https://github.com/kriegaex/Maven_Surefire_PrintToConsoleProblems.
