+++ surefire-its/src/test/resources/surefire-1881-java-agent/pom.xml
@@ -4,7 +4,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <groupId>org.apache.maven.plugins.surefire</groupId>
-    <artifactId>surefire-1881</artifactId>
+    <artifactId>surefire-1881-java-agent</artifactId>
     <version>1.0-SNAPSHOT</version>
 
     <properties>
@@ -98,7 +98,7 @@
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
-            <version>4.13.1</version>
+            <version>4.13.2</version>
             <scope>test</scope>
         </dependency>
     </dependencies>
