+++ surefire-its/src/test/resources/surefire-1881-java-agent/src/test/java/de/scrum_master/dummy/AgentIT.java
