+++ surefire-its/src/test/resources/surefire-1881-jvm-logging/src/test/java/de/scrum_master/dummy/MyTest.java
@@ -1,0 +1,14 @@
+package de.scrum_master.dummy;
+
+import org.junit.Test;
+
+import static org.junit.Assert.assertTrue;
+
+public class MyTest {
+    @Test
+    public void test() {
+        System.out.println("[Test OUT] Hello Maven!");
+        System.err.println("[Test ERR] Hello Maven!");
+        assertTrue(true);
+    }
+}
